+++ providers/index.tsx
@@ -13,15 +13,23 @@ import {
 } from '@/utils/localstorage';
 import { AuthProvider, setAccessCheckResponse, TenantProvider } from '@iblai/iblai-js/web-utils';
 import { getTenant, getUserName, hasNonExpiredAuthToken, redirectToAuthSpa } from '@/utils/helpers';
-import { usePathname } from 'next/navigation';
+import { usePathname, useRouter } from 'next/navigation';
 import { updateRbacPermissions } from '@/features/rbac';
+import { selectRequestedTenant } from '@/features/tenant';
+import { useAppSelector } from '@/lib/hooks';
+import { Spinner } from '@/components/spinner';
 export default function Providers({ children }: { children: React.ReactNode }) {
   const pathname = usePathname();
+  const router = useRouter();
   const [ready, setReady] = useState(false);
   const { saveCurrentTenant } = useCurrentTenant();
   const { saveUserTenants } = useUserTenants();
+  const requestedTenant = useAppSelector(selectRequestedTenant);
   const isSsoLoginRoute = /^\/sso-login/.test(pathname);
   const isVersionRoute = /^\/version/.test(pathname);
+
+  console.log('################### [Providers] isSsoLoginRoute', isSsoLoginRoute);
+  console.log('################### [Providers] isVersionRoute', isVersionRoute);
 
   const loadDataLayer = () => {
     initializeDataLayer(
@@ -62,6 +70,8 @@ export default function Providers({ children }: { children: React.ReactNode }) {
 
     map.set(new RegExp('^/sso-login'), async () => false);
 
+    map.set(new RegExp('^/sso-login-complete'), async () => false);
+
     // allow user to go to version page without auth
     map.set(new RegExp('^\/version'), async () => false);
 
@@ -92,16 +102,23 @@ export default function Providers({ children }: { children: React.ReactNode }) {
       <TenantProvider
         skip={isSsoLoginRoute || isVersionRoute}
         currentTenant={getTenant() || ''}
-        requestedTenant={getTenant() || ''}
+        requestedTenant={requestedTenant}
         saveCurrentTenant={saveCurrentTenant}
         saveUserTenants={saveUserTenants}
-        handleTenantSwitch={handleTenantSwitch}
+        handleTenantSwitch={(tenant, saveRedirect) => handleTenantSwitch(tenant, saveRedirect)}
         username={getUserName() || ''}
         onAuthFailure={(reason) => {
           console.error('[TenantProvider] Auth failure:', reason);
-          window.location.href = '/error/403';
+          router.push('/error/403');
         }}
         onLoadPlatformPermissions={onLoadPlatformpermissions}
+        fallback={
+          <div className="flex h-dvh w-screen items-center justify-center">
+            <div className="space-y-3">
+              <Spinner className="h-14 w-14 text-amber-500" />
+            </div>
+          </div>
+        }
       >
         {children}
       </TenantProvider>